requirements.txt
git+git://github.com/aiogram/aiogram.git@82b1b1ab03c7b958a41b305259654b94a820aa7d
pydantic
pydantic[dotenv]
emoji == 1.4.2
asyncpg == 0.24.0
loguru == 0.5.3
aiogram-calendar == 0.1.1
fastapi == 0.68.1